moved categorical/markets selectors to the categories/markets folders
