+++ src/modules/app/containers/markets-inner-nav.js
@@ -2,23 +2,23 @@ import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
 import { compose } from "redux";
-import { QUERY_VALUE_DELIMITER } from "src/modules/routes/constants/query-value-delimiter";
+import { QUERY_VALUE_DELIMITER } from "modules/routes/constants/query-value-delimiter";
 import {
   CATEGORY_PARAM_NAME,
   TAGS_PARAM_NAME,
   FILTER_SEARCH_PARAM
-} from "src/modules/filter-sort/constants/param-names";
-import { selectCategories } from "src/modules/categories/selectors/categories";
+} from "modules/filter-sort/constants/param-names";
+import { selectCategories } from "modules/categories/selectors/categories";
 import { selectIsMobile } from "src/select-state";
 
 import { isEmpty, map, getOr } from "lodash/fp";
-import { curriedToggleMemberOfArray } from "src/utils/toggle-member-of-array";
-import makeQuery from "src/modules/routes/helpers/make-query";
-import makePath from "src/modules/routes/helpers/make-path";
-import { MARKETS } from "src/modules/routes/constants/views";
-import BaseInnerNavPure from "src/modules/app/components/inner-nav/base-inner-nav-pure";
-import { selectAllCategories } from "src/modules/app/selectors/select-all-categories";
-import { getSelectedTagsAndCategoriesFromLocation } from "src/modules/markets/helpers/get-selected-tags-and-categories-from-location";
+import { curriedToggleMemberOfArray } from "utils/toggle-member-of-array";
+import makeQuery from "modules/routes/helpers/make-query";
+import makePath from "modules/routes/helpers/make-path";
+import { MARKETS } from "modules/routes/constants/views";
+import BaseInnerNavPure from "modules/app/components/inner-nav/base-inner-nav-pure";
+import { selectAllCategories } from "modules/categories/selectors/select-all-categories";
+import { getSelectedTagsAndCategoriesFromLocation } from "modules/markets/helpers/get-selected-tags-and-categories-from-location";
 import noop from "src/utils/noop";
 
 const mapStateToProps = (
