+++ src/modules/categories/selectors/select-all-categories.js
